backend revision first day done

+++ src/main/java/com/example/SpringLogin/Services/AdminService/AffectationModuleService.java
@@ -43,33 +43,46 @@ public class AffectationModuleService {
         return (Administrateur) contextHandlerClass.getCurrentLoggedInUser().getUtilisateur();
     }
 
+    private boolean thisAdminHasMorePrivilege (Administrateur administrateur){
+        return getAdmin().getPrivilege() > administrateur.getPrivilege();
+    }
+
     public AffectationModule makeAffectation(AffectationModule affectationModule) throws Exception {
+        Module module = adminModuleService.getModuleByID(
+                affectationModule.getAffectationModuleId().getModuleId()
+        );
+        Enseignant enseignant = adminEnseignantService.getEnseignantByID(
+                affectationModule.getAffectationModuleId().getEnseignantId()
+        );
+        AffectationModule existedSameAffectation =  affectationModuleRepo.findByEnseignantAndModule(enseignant,module).get();
+        if(existedSameAffectation == null) {
+            if (affectationModule.getType() == teachingType.COURSE) {
+                if( !hasNotAffectedForCoursesYet(module)) {
+                    throw new Exception("this module already affected to a teacher if you are sure you want that please start by deleting the existing affectation first then reAffect it ");
+                }
+            } else{
+                if(!module.isHasTDTP())
+                    throw new Exception("this module don't have a TD_TP session you can't make this affectation");
+            }
 
-        if( affectationModule.getType() == teachingType.COURSE
-           && !hasNotAffectedForCoursesYet(affectationModule.getModule())
-       ){
-           throw new Exception("this module already affected to a teacher if you are sure you want that please start by deleting the affectation first then reAffect it ");
-       }
-        else{
-//            Administrateur administrateur;
-//            administrateur = administrateurRepo.findById(1L).get();
-//            affectationModule.setAdmin(administrateur);
-//            Module module = entityManager.getReference(Module.class, affectationModule.getAffectationModuleId().getModuleId());
-//            Enseignant enseignant = entityManager.getReference(Enseignant.class, affectationModule.getAffectationModuleId().getEnseignantId());
-            Module module = adminModuleService.getModuleByID(
-                    affectationModule.getAffectationModuleId().getModuleId()
-            );
-            Enseignant enseignant = adminEnseignantService.getEnseignantByID(
-                    affectationModule.getAffectationModuleId().getEnseignantId()
-            );
-
-            affectationModule.setModule(module);
-            affectationModule.setEnseignant(enseignant);
-            affectationModule.setAffectationDate(Timestamp.from(Instant.now()));
-            affectationModule.setAdmin(getAdmin());
-            //entityManager.persist(affectationModule);
-            return affectationModuleRepo.saveAndFlush(affectationModule);
         }
+        else{
+            if(existedSameAffectation.getType() == affectationModule.getType())
+                throw new Exception("this affectation already made ");
+
+            if(affectationModule.getType() == teachingType.COURSE){
+                if(thisAdminHasMorePrivilege(existedSameAffectation.getAdmin())){
+                    throw new Exception("this teacher is affected as TD_TP for this module to upgrade him for COURSE you have to delete first the TD_TP affectation which you don't have the right privilege for it ");
+                }
+                affectationModuleRepo.deleteById(existedSameAffectation.getAffectationModuleId());
+            }else{
+                throw new Exception("this teacher is affected for courses no need for TD_TP all the functionalities possible when the type is course  ");
+            }
+        }
+
+        affectationModule.setAffectationDate(Timestamp.from(Instant.now()));
+        affectationModule.setAdmin(getAdmin());
+        return affectationModuleRepo.saveAndFlush(affectationModule);
     }
 
     private boolean hasNotAffectedForCoursesYet(Module module){
@@ -80,7 +93,11 @@ public class AffectationModuleService {
             return false;
     }
 
-    public void deleteAffectation(Module module, Enseignant enseignant){
+    public void deleteAffectation(Module module, Enseignant enseignant) throws Exception {
+        AffectationModule existedAffectation = affectationModuleRepo.findByEnseignantAndModule(enseignant,module).get();
+        if(thisAdminHasMorePrivilege(existedAffectation.getAdmin())){
+            throw new Exception("you don't have the right privilege for this action");
+        }
         affectationModuleRepo.deleteByModuleAndEnseignant(module, enseignant);
     }
